Migrate controller + reconciler to new file, new format, TODO logging commented
